requirements.txt
apify-client>=1.6.0
openai>=1.30.0
gspread>=6.0.0
google-auth>=2.29.0
python-dotenv>=1.0.0
